+++ ruby/red-arrow/ext/arrow/record-batch.cc
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-#include "red_arrow.hpp"
+#include "red-arrow.hpp"
 
 #include <ruby.hpp>
